requirements.txt
discord.py==2.4.0
PyNaCl
aiohttp

